contrib/fuel-utils/requirements.txt
paramiko>=1.12
eventlet>=0.12
greenlet>=0.3.1
python-neutronclient
python-keystoneclient
